Move response classes into their own files
