+++ src/SourceControl/AzureDevOps/POCOs/ContentMetadata.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SourceControl.AzureDevOps.POCOs
+{
+    public class ContentMetadata
+    {
+        public string fileName { get; set; }
+    }
+}
